+++ src/components/layout/mdl-layout-drawer.component.ts
@@ -25,7 +25,4 @@ export class MdlLayoutDrawerComponent {
 
   }
 
-  public isDrawerDirectChildOf(layout: MdlLayoutComponent) {
-    return this.parentLayout === layout;
-  }
 }