+++ src/components/table/mdl-table.component.ts
@@ -2,8 +2,7 @@ import {
   Component,
   Input,
   Output,
-  EventEmitter,
-  ComponentMetadata
+  EventEmitter
 } from '@angular/core';
 import {
   IMdlTableModel,