+++ src/components/textfield/mdl-textfield.component.spec.ts
@@ -349,6 +349,32 @@ describe('Component: MdlTextField', () => {
     expect(component.onBlur).toHaveBeenCalled();
   });
 
+  it('should be possible to set the focus programmatically', () => {
+    TestBed.overrideComponent(MdlTestComponent, { set: {
+      template: `
+          <mdl-textfield  type="text"></mdl-textfield>'
+      `
+    }});
+    let fixture = TestBed.createComponent(MdlTestComponent);
+
+    let textFieldDebugElement = fixture.debugElement.query(By.directive(MdlTextFieldComponent));
+    let textFieldComonent = textFieldDebugElement.componentInstance;
+    let el = textFieldDebugElement.nativeElement;
+
+    // if called here the inputEl is not set - if the setFocus didn't check for inoutEl the next line throws.
+    textFieldComonent.setFocus();
+
+
+    fixture.detectChanges();
+
+    // now it is save to call setFocus and the focus is set.
+    textFieldComonent.setFocus();
+
+    fixture.detectChanges();
+
+    expect(el.classList.contains('is-focused')).toBe(true);
+
+  });
 
 });